Load the archive date cutoff from the environment
Instead of having this setting in `configuration.yaml` (forcing anyone
wanting to use it on their own system to edit the file and redeploy), we
can read it in from the environment.

In the case where the environment variable is not set, I'm taking that
to mean that stale channels can be archived no matter the current date.
Given the old value set in the config file is now in the past, this
effectively doesn't change the behaviour.

I've made a few changes at the point this config value is used, so that
it gets parsed at the point of construction rather than the point of
use. This was mostly for clean-up purposes, but has the added benefit of
causing the code to error out earlier if it's given an invalid value.

There's one point where this date gets passed directly into a `format`
call, but that produces the same value as before.