Improve default handling in archive date code
By specifying a default value before parsing, rather than after, it's
possible to remove the conditional and always parse *something*, even a
default value.

This uses `or` rather than the two-argument call to `getenv` so that an
empty value for the environment variable is treated the same as an unset
value. If this code were to use `getenv` with a specified default,
running with the `EARLIEST_ARCHIVE_DATE` envvar set to an empty string
would cause the program to raise an exception, rather than fall back to
the default.