Ensure logger handlers aren't doubled up
Bail on setting up a logger if it already has handlers added.

This happens because `logging.getLogger()` returns a singleton
instance of a `Logger` when provided with the same name. So we're
picking up the same logger across multiple files and feedin it
into this `set_up_logger()` method.

Now the method is smart enough not to step on its past selves.