Remove slack logging altogether
Once we combined logging to both slack and stdout using the shared logger, it made it so that either slack logging was rate limited or the log level was too high for useful information. This removes it, but we can add it back in in select places using the slackbot directly.